requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
beautifulsoup4==4.12.3
playwright==1.44.0
psycopg2-binary